# backend/docs/swagger.yaml
basePath: /api/v1
definitions:
...
  model.DeviceCommandRequest:
    properties:
      b:
        type: integer
      brightness:
        type: integer
      color_temp_mireds:
        type: integer
      g:
        type: integer
      hue:
        type: integer
      "on":
        type: boolean
      r:
        type: integer
      saturation:
        type: integer
      scene_id:
        type: integer
    type: object
  model.DeviceRegisterRequest:
    properties:
      ip_address:
        type: string
      name:
        type: string
      room_id:
        type: string
    type: object
  model.DeviceResponse:
    properties:
      commissioned_at:
        type: string
      current_state:
        additionalProperties: {}
        type: object
      device_type:
        type: string
      id:
        type: string
      ip_address:
        type: string
      last_seen:
        type: string
      mac_address:
        type: string
      manufacturer:
        type: string
      model:
        type: string
      name:
        type: string
      room_id:
        type: string
      status:
        type: string
    type: object
...
host: localhost:8000
info:
  contact: {}
  description: Backend REST API for Jeffries Homeapp.
  title: Jeffries Homeapp API
  version: "1.0"
paths:
...
  /devices:
    get:
      description: Returns a list of all devices in the homeapp
      parameters:
      - default: 0
        description: Skip count
        in: query
        name: skip
        type: integer
      - default: 100
        description: Limit count
        in: query
        name: limit
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/model.DeviceResponse'
            type: array
        "500":
          description: Failed to fetch devices
          schema:
            type: string
      summary: List all devices
      tags:
      - Devices
  /devices/{deviceID}:
    delete:
      description: Deletes a device by its ID
      parameters:
      - description: Device ID (UUID)
        in: path
        name: deviceID
        required: true
        type: string
      responses:
        "204":
          description: No Content
        "400":
          description: Invalid device ID
          schema:
            type: string
        "404":
          description: Device not found
          schema:
            type: string
        "500":
          description: Internal Server Error
          schema:
            type: string
      summary: Delete a device
      tags:
      - Devices
    get:
      description: Returns a single device by its ID
      parameters:
      - description: Device ID (UUID)
        in: path
        name: deviceID
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/model.DeviceResponse'
        "400":
          description: Invalid device ID
          schema:
            type: string
        "404":
          description: Device not found
          schema:
            type: string
        "500":
          description: Internal Server Error
          schema:
            type: string
      summary: Get a device
      tags:
      - Devices
    patch:
      consumes:
      - application/json
      description: Modifies an existing device
      parameters:
      - description: Device ID (UUID)
        in: path
        name: deviceID
        required: true
        type: string
      - description: Updated Device Details
        in: body
        name: request
        required: true
        schema:
          additionalProperties: true
          type: object
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/model.DeviceResponse'
        "400":
          description: Invalid request body or ID
          schema:
            type: string
        "404":
          description: Device not found
          schema:
            type: string
        "500":
          description: Internal Server Error
          schema:
            type: string
        "502":
          description: WiZ lamp unreachable
          schema:
            type: string
      summary: Update a device
      tags:
      - Devices
  /devices/{deviceID}/command:
    post:
      consumes:
      - application/json
      description: Controls a WiZ bulb via UDP (on/off, brightness, color, scene)
      parameters:
      - description: Device ID (UUID)
        in: path
        name: deviceID
        required: true
        type: string
      - description: Command Details
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/model.DeviceCommandRequest'
      produces:
      - application/json
      responses:
        "204":
          description: No Content
        "400":
          description: Invalid request body or ID
          schema:
            type: string
        "404":
          description: Device not found
          schema:
            type: string
        "422":
          description: Device has no IP address
          schema:
            type: string
        "502":
          description: WiZ command failed
          schema:
            type: string
      summary: Send command to WiZ device
      tags:
      - Devices
  /devices/register:
    post:
      consumes:
      - application/json
      description: Registers a new WiZ bulb via local UDP verification
      parameters:
      - description: Device Registration Details
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/model.DeviceRegisterRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/model.DeviceResponse'
        "400":
          description: Invalid request body or missing IP/name
          schema:
            type: string
        "500":
          description: Internal Server Error
          schema:
            type: string
        "502":
          description: Cannot reach WiZ bulb
          schema:
            type: string
      summary: Register a WiZ device
      tags:
      - Devices
...
securityDefinitions:
  ApiKeyAuth:
    in: header
    name: X-API-Key
    type: apiKey
swagger: "2.0"